Revert "Merge pull request #7430 from vector-im/andy/user_trust"
This reverts commit 7f5d3b46eb02a7801d4a7c781762eea3d16ee0e6, reversing
changes made to 5a3adde8f70daec795ac1bee17c5159a155c6106.
