+++ Riot/Modules/EncryptionInfo/EncryptionInfoView.h
