+++ Riot/SupportingFiles/Riot-Bridging-Header.h
@@ -18,7 +18,6 @@
 #import "RoomBubbleCellData.h"
 #import "MXKRoomBubbleTableViewCell+Riot.h"
 #import "UserEncryptionTrustLevel.h"
-#import "RoomEncryptionTrustLevel.h"
 #import "RoomReactionsViewSizer.h"
 #import "RoomEncryptedDataBubbleCell.h"
 #import "LegacyAppDelegate.h"
