+++ Riot/Utils/EncryptionTrustLevelBadgeImageHelper.swift
