+++ RiotShareExtension/SupportingFiles/RiotShareExtension-Bridging-Header.h
@@ -6,8 +6,6 @@
 #import "AvatarGenerator.h"
 #import "BuildInfo.h"
 #import "ShareItemSender.h"
-#import "UserEncryptionTrustLevel.h"
-#import "RoomEncryptionTrustLevel.h"
 
 // MatrixKit imports
 #import "MatrixKit-Bridging-Header.h"
