+++ RiotShareExtension/target.yml
@@ -87,4 +87,3 @@ targets:
         - "**/*.md" # excludes all files with the .md extension
     - path: ../Riot/Modules/Room/TimelineCells/Styles/RoomTimelineStyleIdentifier.swift
     - path: ../Riot/Modules/VoiceBroadcast/VoiceBroadcastSDK/MatrixSDK
-    - path: ../Riot/Modules/Encryption/EncryptionTrustLevel.swift
